azure-functions/requirements.txt
azure-functions
numpy>=1.21.0
scikit-learn>=1.0.0
azure-identity>=1.12.0
azure-digitaltwins-core>=1.2.0
azure-core>=1.24.0
